Phase 7B Continued: Major warning reduction in critical files
Significant progress on systematic warning reduction:

Files Fixed:
- unitsX.c: 8 → 0 warnings (100% reduction - CLEAN!)
- enlistG.c: 26 → 6 warnings (77% reduction)
- ioG.c: 44 → 26 warnings (41% reduction)

Priority 1 Format Specifier Fixes:
- enlistG.c: Fixed %ld/%d mismatches with literal 1 and max_numunit() returns
- Corrected long int format specifiers throughout

Priority 2 Type Safety Fixes:
- unitsX.c: Fixed enum Aclasstype vs int comparisons with explicit casts
- ioG.c: Fixed size_t vs int comparisons in screen coordinate calculations
- Enhanced type safety across all affected comparison operations

Technical Impact:
- Eliminated critical format string vulnerabilities
- Resolved signed/unsigned comparison warnings
- Improved C2023 standards compliance
- One file now compiles completely warning-free

+++ Src/ioG.c
@@ -436,7 +436,7 @@ enter_unittype PARM_3(int, style, char *, prompt, char *, emsg)
       if (!a_ismerc(count)) continue;
       break;
     case 2:
-      if (ainfo_list[count].class != upg_class) continue;
+      if ((int)ainfo_list[count].class != upg_class) continue;
       break;
     case 3:
       if (!a_enlistaway(count)) continue;
@@ -464,7 +464,7 @@ enter_unittype PARM_3(int, style, char *, prompt, char *, emsg)
       }
 
       /* show as a choice */
-      if (xloc + strlen(ainfo_list[count].select) > COLS - 3) {
+      if (xloc + (int)strlen(ainfo_list[count].select) > COLS - 3) {
 	xloc = 1;
 	yloc++;
 	move(yloc, xloc);
@@ -521,7 +521,7 @@ enter_unittype PARM_3(int, style, char *, prompt, char *, emsg)
       if (!a_ismerc(count)) continue;
       break;
     case 2:
-      if (ainfo_list[count].class != upg_class) continue;
+      if ((int)ainfo_list[count].class != upg_class) continue;
       break;
     case 3:
       if (!a_enlistaway(count)) continue;
@@ -598,7 +598,7 @@ get_dmode PARM_1(char *, prompt)
     }
 
     /* now check it */
-    if (xloc + strlen(dmode_tptr->d.name) > COLS - 2) {
+    if (xloc + (int)strlen(dmode_tptr->d.name) > COLS - 2) {
       if (yloc == LINES - 3) {
 	mvaddstr(yloc, xloc, "...");
 	break;
@@ -892,7 +892,7 @@ get_diplomacy PARM_2(int, oldstatus, int, othstatus)
     } else {
       strcpy(buffer, dipname[i]);
     }
-    if (xloc + strlen(buffer) > COLS - 5) {
+    if (xloc + (int)strlen(buffer) > COLS - 5) {
       yloc++;
       xloc = 2;
     }
